+++ .gitignore
@@ -1,3 +1,5 @@
 settings.json
 
+test.js
+
 .github/
