--- JS/module/Auth/userData.js
@@ -1,13 +1,0 @@
-
-const USER_DATA = {
-  id1 : {
-    Email : 'test@codeit.kr',
-    password : 'codeit101'
-  },
-  id2 : {
-    Email: 'nahyun@codeit.kr',
-    password : '123456'
-  }
-}
-
-export  { USER_DATA }
