+++ JS/module/Auth/validation.js
@@ -1,57 +1,62 @@
-import { USER_INPUT_ID, requestSettings } from "./requestSettings.js";
+import { inputEmail, errorEmail, inputPassword, errorPassword, inputPassword2, errorPassword2} from "./variables.js";
+import { removeErrorStyle, errorStyle, EMAIL_EMPTY,EMAIL_NOT_VALID,EMAIL_OCCUPIED,PASSWORD_EMPTY,PASSWORD_NOT_SAME,PASSWORD_NOT_VALID } from "./errorStyle.js";
+import { checkFormEmpty, checkEmailValid, checkPasswordValid,checkStringSame,checkEmailAvailable} from "./conditions.js";
 
-// 빈 칸인지 검사하는 함수
-
-export function checkFormEmpty(inputTag) {
-  const isNotEmpty = !(inputTag.length === 0);
-  return isNotEmpty ? true : false;
-}
-
-export function checkEmailValid(email) {
-  const EMAIL_REGEX = new RegExp(
-    /^[0-9a-zA-Z]([-_\.]?[0-9a-zA-Z])*@[0-9a-zA-Z]([-_\.]?[0-9a-zA-Z])*\.[a-zA-Z]{2,3}$/i
-  );
-  const isValid = EMAIL_REGEX.test(email.trim());
-  return isValid ? true : false;
-}
-
-export function checkPasswordValid(password) {
-  const PASSWORD_REGEX = new RegExp(
-    /^(?=.*\d)(?=.*[a-zA-Z])[0-9a-zA-Z]{8,16}$/
-  );
-  const isValid = PASSWORD_REGEX.test(password.trim());
-  return isValid ? true : false;
-}
-
-export function checkStringSame(value1, value2) {
-  return value1 === value2 ? true : false;
-}
-
-export function checkErrorMessagesExist() {
-  let isAllValid = true;
-  const errorMessages = document.querySelectorAll(".error_message");
-  for (let errorTag of errorMessages) {
-    if ([...errorTag.classList].includes("show")) {
-      isAllValid = false;
-    }
-  }
-  return isAllValid;
-}
-
-export async function checkEmailAvailable() {
-  try {
-    const ID = USER_INPUT_ID();
-    const response = await fetch(
-      "https://bootcamp-api.codeit.kr/api/check-email",
-      requestSettings("POST", ID)
-    );
-    const result = await response.json();
-    if (response.status === 200) {
-      return true;
-    } else if (response.status === 409) {
-      return false;
-    }
-  } catch (error) {
-    return false;
+export function emailValidationSignIn(){
+  const email = inputEmail.value;
+  if (!checkFormEmpty(email))
+    return errorStyle(inputEmail, errorEmail,EMAIL_EMPTY)
+  else if (!checkEmailValid(email)) {
+    return errorStyle(inputEmail, errorEmail, EMAIL_NOT_VALID)
+  } else {
+    return true
   }
 }
+
+export function passwordValidationSignIn(){
+  const password = inputPassword.value;
+  if (!checkFormEmpty(password)){
+  errorStyle(inputPassword, errorPassword, PASSWORD_EMPTY)
+  return false
+  }
+  return true
+}
+
+export async function emailValidationSignUp() {
+  const email = inputEmail.value;
+  if (!checkFormEmpty(email)) {
+    return errorStyle(inputEmail,errorEmail,EMAIL_EMPTY);
+  } else if (!checkEmailValid(email)) {
+    return errorStyle(inputEmail,errorEmail,EMAIL_NOT_VALID);
+  } else if (!await checkEmailAvailable()) {
+    return errorStyle(inputEmail,errorEmail,EMAIL_OCCUPIED);
+  } else {
+    removeErrorStyle(inputEmail, errorEmail);
+    return true;
+  }
+}
+
+export function passwordValidationSignUp() {
+  const password = inputPassword.value;
+  if (!checkFormEmpty(password)) {
+    return errorStyle(inputPassword,errorPassword,PASSWORD_EMPTY);
+  }
+  if (!checkPasswordValid(password)) {
+    return errorStyle(inputPassword,errorPassword,PASSWORD_NOT_VALID);
+  }
+  removeErrorStyle(inputPassword, errorPassword);
+  return true;
+}
+
+export function password2ValidationSignUp() {
+  const password = inputPassword.value;
+  const password2 = inputPassword2.value;
+  if (!checkFormEmpty(password2)) {
+    return errorStyle(inputPassword2,errorPassword2,PASSWORD_EMPTY);
+  } else if (!checkStringSame(password, password2)) {
+    return errorStyle(inputPassword2,errorPassword2,PASSWORD_NOT_SAME);
+  }
+  removeErrorStyle(inputPassword2, errorPassword2);
+  return true;
+}
+
